fix: update podcast action options to a separate component
This was made to avoid IPC requests to tauri events, the useDownload hook setup the listener to the tauri event that tells the front end when downloads finishes. Now this event listener is only dispatched when the actions dropdown is spawned.
