+++ src/app/hooks/use-podcast-options.tsx
@@ -4,6 +4,8 @@ import { ROUTES } from '@/routes/routesList'
 import { podcasts } from '@/service/podcasts'
 import { Episode } from '@/types/responses/podcasts'
 import { queryKeys } from '@/utils/queryKeys'
+import { isTauri } from '@/utils/tauriTools'
+import { useDownload } from './use-download'
 
 interface PodcastOptionsProps {
   episode: Episode
@@ -12,6 +14,15 @@ interface PodcastOptionsProps {
 export function usePodcastOptions({ episode }: PodcastOptionsProps) {
   const navigate = useNavigate()
   const queryClient = useQueryClient()
+  const { downloadBrowser, downloadTauri } = useDownload()
+
+  function handleDownload() {
+    if (isTauri()) {
+      downloadTauri(episode.audio_url, episode.id)
+    } else {
+      downloadBrowser(episode.audio_url, episode.id)
+    }
+  }
 
   const markAsPlayedMutation = useMutation({
     mutationFn: () =>
@@ -37,6 +48,7 @@ export function usePodcastOptions({ episode }: PodcastOptionsProps) {
   }
 
   return {
+    handleDownload,
     markAsPlayedMutation,
     gotoPodcast,
     gotoEpisode,
